Rename members to users to reflect service name
