+++ services/services.go
@@ -1,8 +1,8 @@
 /**
 * Copyright © 2017, ACM@UIUC
 *
-* This file is part of the Groot Project.  
-* 
+* This file is part of the Groot Project.
+*
 * The Groot Project is open source software, released under the University of
 * Illinois/NCSA Open Source License. You should have received a copy of
 * this license in a file with the distribution.
@@ -40,7 +40,7 @@ func Index(w http.ResponseWriter, r *http.Request) {
 func RegisterAPIs() {
 	Routes = append(Routes, AuthRoutes...)
 	Routes = append(Routes, GroupsRoutes...)
-	Routes = append(Routes, MembersRoutes...)
+	Routes = append(Routes, UsersRoutes...)
 	Routes = append(Routes, QuotesRoutes...)
 	Routes = append(Routes, RecruitersRoutes...)
 	Routes = append(Routes, HardwareRoutes...)
